Progress on Simulator GUI class
