+++ src/main/java/com/ezasm/Arguments.java
@@ -1,8 +1,9 @@
 package com.ezasm;
 
 import com.ezasm.gui.Window;
+import com.ezasm.simulation.ISimulator;
 import com.ezasm.simulation.Memory;
-import com.ezasm.simulation.Simulator;
+import com.ezasm.simulation.NewSimulator;
 import org.apache.commons.cli.*;
 
 import javax.swing.*;
@@ -72,14 +73,14 @@ public class Arguments {
             wordSize = Memory.DEFAULT_WORD_SIZE;
         }
 
-        Simulator sim = new Simulator(wordSize, memorySize, config.getSimSpeed());
+        ISimulator sim = new NewSimulator(wordSize, memorySize);
         String filepath = "";
         if (commandLine.hasOption(fileOption)) {
             filepath = commandLine.getOptionValue(fileOption);
         }
 
         if (commandLine.hasOption(windowlessOption)) {
-            CommandLineInterface cli = null;
+            CommandLineInterface cli;
             if (filepath.equals("")) {
                 cli = new CommandLineInterface(sim);
             } else {
