+++ src/main/java/com/ezasm/gui/MenubarFactory.java
@@ -47,6 +47,8 @@ public class MenubarFactory {
         menu.addSeparator();
         addMenuItem(menu, EXIT);
 
+
+
         /*
          * SETTINGS
          */
